pyproject.toml
[project]
name = "dash-estoque"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "pandas>=3.0.2",
    "plotly-express>=0.4.1",
    "streamlit>=1.56.0",
]
